Перенесены все функции со стилем из отдельного файла в основной файл программы, по заданию ревьюера. Починены шрифты. отключена возможность нажатия на кнопку до появления следующего вопроса.
