+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -1,6 +1,6 @@
 import UIKit
 
-struct QuizQuestion{
+struct QuizQuestion {
     //название картинки(фильма) из assets
     let image: String
     //строка с вопросом о рейтинге фильма
@@ -20,16 +20,17 @@ struct QuizStepViewModel {
 
 // для состояния "Результат квиза"
 struct QuizResultsViewModel {
-  // строка с заголовком алерта
-  let title: String
-  // строка с текстом о количестве набранных очков
-  let text: String
-  // текст для кнопки алерта
-  let buttonText: String
+    // строка с заголовком алерта
+    let title: String
+    // строка с текстом о количестве набранных очков
+    let text: String
+    // текст для кнопки алерта
+    let buttonText: String
 }
 
 
-
+//ДЛЯ РЕВЬЮЕРА!!//
+// пожалуйста отразите в ревью работает ли шрифт, так как у меня попался баг с добавлением шрифтов, я сделал все как указанно в иструкции повторял несколько раз, искал решение в интернете и у наставника, наставник посаветовал шрифт прописать вручную, собственно так я и сделал.
 final class MovieQuizViewController: UIViewController {
     
     private var currentQuestionIndex = 0
@@ -73,12 +74,12 @@ final class MovieQuizViewController: UIViewController {
     
     // MARK: - Lifecycle
     
-    @IBOutlet weak var questionTitleLabel: UILabel!
-    @IBOutlet weak var counterLabel: UILabel!
-    @IBOutlet weak var imageView: UIImageView!
-    @IBOutlet weak var questionLabel: UILabel!
-    @IBOutlet weak var yesButton: UIButton!
-    @IBOutlet weak var noButton: UIButton!
+    @IBOutlet private weak var questionTitleLabel: UILabel!
+    @IBOutlet private weak var counterLabel: UILabel!
+    @IBOutlet private weak var imageView: UIImageView!
+    @IBOutlet private weak var questionLabel: UILabel!
+    @IBOutlet private weak var yesButton: UIButton!
+    @IBOutlet private weak var noButton: UIButton!
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -87,16 +88,19 @@ final class MovieQuizViewController: UIViewController {
     }
     
     @IBAction private func noButtonPressed(_ sender: UIButton) {
+        buttonsIsDisabled()
         let userAnswer = questions[currentQuestionIndex].correctAnswer == false
         showAnswerResult(isCorrect: userAnswer)
     }
     
     @IBAction private func yesButtonPressed(_ sender: UIButton) {
+        buttonsIsDisabled()
         let userAnswer = questions[currentQuestionIndex].correctAnswer == true
         showAnswerResult(isCorrect: userAnswer)
-      
+        
     }
-    private func showAnswerResult(isCorrect: Bool){
+    
+    private func showAnswerResult(isCorrect: Bool) {
         if isCorrect == true{
             correctAnswers += 1
         }
@@ -104,14 +108,25 @@ final class MovieQuizViewController: UIViewController {
         DispatchQueue.main.asyncAfter(deadline: .now() + 1) {
             self.imageView.layer.borderWidth = 0
             self.showNextQuestionOrResult()
+            self.buttonsIsEnabled()
         }
     }
     
-    private func showNextQuestionOrResult(){
+    private func buttonsIsDisabled(){
+        noButton.isEnabled = false
+        yesButton.isEnabled = false
+    }
+    
+    private func buttonsIsEnabled(){
+        noButton.isEnabled = true
+        yesButton.isEnabled = true
+    }
+    
+    private func showNextQuestionOrResult() {
         if currentQuestionIndex == questions.count - 1 {
             let result = QuizResultsViewModel(title: "Этот раунд окончен!",
-                                                      text: "ваш результат \(correctAnswers)/\(questions.count)",
-                                                      buttonText: "Сыграть еще раз?")
+                                              text: "ваш результат \(correctAnswers)/\(questions.count)",
+                                              buttonText: "Сыграть еще раз?")
             show(quiz: result)
         }else{
             currentQuestionIndex += 1
@@ -119,7 +134,7 @@ final class MovieQuizViewController: UIViewController {
         }
     }
     
-    private func show(quiz result: QuizResultsViewModel){
+    private func show(quiz result: QuizResultsViewModel) {
         
         let alert = UIAlertController(title: result.title, message: result.text, preferredStyle: .alert)
         let action = UIAlertAction(title: result.buttonText, style: .default){_ in
@@ -131,7 +146,7 @@ final class MovieQuizViewController: UIViewController {
         self.present(alert, animated: true)
     }
     
-    private func show(quiz step: QuizStepViewModel){
+    private func show(quiz step: QuizStepViewModel) {
         imageView.image = step.image
         questionLabel.text = step.question
         counterLabel.text = step.questionNumber
@@ -141,6 +156,67 @@ final class MovieQuizViewController: UIViewController {
         QuizStepViewModel(image: UIImage(named: model.image) ?? UIImage(),
                           question: model.text,
                           questionNumber: "\(currentQuestionIndex + 1)/\(questions.count)")
+    }
+    
+    func screenStyle() {
+        viewStyle()
+        questionTitleLabelStyle()
+        counterLabelStyle()
+        imageViewStyle()
+        questionLabelStyle()
+        yesButtonStyle()
+        noButtonStyle()
+    }
+    
+    func viewStyle(){
+        view.backgroundColor = .ypBlack
+    }
+    
+    func questionTitleLabelStyle() {
+        questionTitleLabel.font = UIFont(name: "YSDisplay-Medium", size: 20)
+        questionTitleLabel.textColor = .ypWhite
+    }
+    
+    func counterLabelStyle() {
+        counterLabel.font = UIFont(name: "YSDisplay-Medium", size: 20)
+        counterLabel.textColor = .ypWhite
+    }
+    
+    func imageViewStyle() {
+        imageView.layer.cornerRadius = 20
+        imageView.contentMode = .scaleAspectFill
+        imageView.backgroundColor = .ypWhite
+    }
+    
+    func questionLabelStyle() {
+        questionLabel.textColor = .ypWhite
+        questionLabel.font = UIFont(name: "YSDisplay-Bold", size: 23)
+        questionLabel.numberOfLines = 2
+        questionLabel.textAlignment = .center
+    }
+    
+    func yesButtonStyle() {
+        yesButton.setTitle("Да", for: .normal)
+        yesButton.titleLabel?.font = UIFont(name: "YSDisplay-Medium", size: 20)
+        yesButton.setTitleColor(.ypBlack, for: .normal)
+        yesButton.layer.cornerRadius = 15
+        yesButton.backgroundColor = .ypWhite
+       
+    }
+    func noButtonStyle() {
+        noButton.setTitle("Нет", for: .normal)
+        noButton.titleLabel?.font = UIFont(name: "YSDisplay-Medium", size: 20)
+        noButton.setTitleColor(.ypBlack, for: .normal)
+        noButton.layer.cornerRadius = 15
+        noButton.backgroundColor = .ypWhite
+        
+    }
+    
+    func imageViewBorderStyle(result isCorrect: Bool) {
+        imageView.layer.masksToBounds = true
+        imageView.layer.borderWidth = 8
+        imageView.layer.borderColor = isCorrect ? UIColor.ypGreen.cgColor : UIColor.ypRed.cgColor
+        imageView.layer.cornerRadius = 20
     }
 }
 
@@ -154,56 +230,56 @@ final class MovieQuizViewController: UIViewController {
  Настоящий рейтинг: 9,2
  Вопрос: Рейтинг этого фильма больше чем 6?
  Ответ: ДА
-
-
+ 
+ 
  Картинка: The Dark Knight
  Настоящий рейтинг: 9
  Вопрос: Рейтинг этого фильма больше чем 6?
  Ответ: ДА
-
-
+ 
+ 
  Картинка: Kill Bill
  Настоящий рейтинг: 8,1
  Вопрос: Рейтинг этого фильма больше чем 6?
  Ответ: ДА
-
-
+ 
+ 
  Картинка: The Avengers
  Настоящий рейтинг: 8
  Вопрос: Рейтинг этого фильма больше чем 6?
  Ответ: ДА
-
-
+ 
+ 
  Картинка: Deadpool
  Настоящий рейтинг: 8
  Вопрос: Рейтинг этого фильма больше чем 6?
  Ответ: ДА
-
-
+ 
+ 
  Картинка: The Green Knight
  Настоящий рейтинг: 6,6
  Вопрос: Рейтинг этого фильма больше чем 6?
  Ответ: ДА
-
-
+ 
+ 
  Картинка: Old
  Настоящий рейтинг: 5,8
  Вопрос: Рейтинг этого фильма больше чем 6?
  Ответ: НЕТ
-
-
+ 
+ 
  Картинка: The Ice Age Adventures of Buck Wild
  Настоящий рейтинг: 4,3
  Вопрос: Рейтинг этого фильма больше чем 6?
  Ответ: НЕТ
-
-
+ 
+ 
  Картинка: Tesla
  Настоящий рейтинг: 5,1
  Вопрос: Рейтинг этого фильма больше чем 6?
  Ответ: НЕТ
-
-
+ 
+ 
  Картинка: Vivarium
  Настоящий рейтинг: 5,8
  Вопрос: Рейтинг этого фильма больше чем 6?
